% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticateWithApp.R
\name{authenticateWithApp}
\alias{authenticateWithApp}
\title{Authenticate Bloomberg API access}
\usage{
authenticateWithApp(con = defaultConnection())
}
\arguments{
\item{con}{A connection object as created by a \code{blpConnect}
call, and retrieved via the internal function
\code{defaultConnection}.}
}
\value{
The returned bloomberg identity object should be passed to subsequent data
calls via bdp(), bds(), etc.
}
\description{
This function authenticates against the the Bloomberg API
}
\examples{
\dontrun{
blpConnect(host=blpHost, port=blpPort, app_name=blpAppName)
blpid <- blpAuthenticate(con)
bdp("IBM US Equity", "NAME", identity=blpid)
}
}
\author{
Alfred Kanzler
}
