% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bdp.R
\name{bdp}
\alias{bdp}
\title{Run 'Bloomberg Data Point' Queries}
\usage{
bdp(con, securities, fields, options = NULL, overrides = NULL,
  identity = NULL)
}
\arguments{
\item{con}{A connection object as returned by a \code{blpConnect} call}

\item{securities}{A character vector with security symbols in
Bloomberg notation.}

\item{fields}{A character vector with Bloomberg query fields.}

\item{options}{An optional named character vector with option
values. Each field must have both a name (designating the option
being set) as well as a value.}

\item{overrides}{An optional named character vector with override
values. Each field must have both a name (designating the override
being set) as well as a value.}

\item{identity}{An optional idendity object.}
}
\value{
A data frame with as a many rows as entries in
\code{securities} and columns as entries in \code{fields}.
}
\description{
This function uses the Bloomberg API to retrieve 'bdp' (Bloomberg
Data Point) queries
}
\examples{
\dontrun{
  con <- blpConnect()   # adjust as needed
  bdp(con, c("ESA Index", "SPY US Equity"), c("PX_LAST", "VOLUME"))
}
}
\author{
Whit Armstrong and Dirk Eddelbuettel
}

