% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bds.R
\name{bds}
\alias{bds}
\title{Run 'Bloomberg Data Set' Queries}
\usage{
bds(con, securities, fields, options = NULL, overrides = NULL,
  identity = NULL)
}
\arguments{
\item{con}{A connection object as returned by a \code{blpConnect} call}

\item{securities}{A character vector with security symbols in
Bloomberg notation.}

\item{fields}{A character string with a single Bloomberg query field.}

\item{options}{An optional named character vector with option
values. Each field must have both a name (designating the option
being set) as well as a value.}

\item{overrides}{An optional override vector.}

\item{identity}{An optional idendity object.}
}
\value{
A list with as a entries as there are entries in
\code{securities}; each list contains a data.frame with one row
per observations and as many columns as entries in
\code{fields}. If the list is of length one, it is collapsed into
a single data frame.
}
\description{
This function uses the Bloomberg API to retrieve 'bds' (Bloomberg
Data Set) queries
}
\examples{
\dontrun{
  con <- blpConnect()   # adjust as needed
  bds(con, "GOOG US Equity", "TOP_20_HOLDERS_PUBLIC_FILINGS")
}
}
\author{
Whit Armstrong and Dirk Eddelbuettel
}

