% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blpAuthenticate.R
\name{blpAuthenticate}
\alias{blpAuthenticate}
\title{Authenticate Bloomberg API access}
\usage{
blpAuthenticate(uuid = getOption("uuid"), host = getOption("blpAuthHost",
  "localhost"), ip.address = getOption("blpIP", NULL),
  con = defaultConnection(), default = TRUE)
}
\arguments{
\item{uuid}{A character variable with a unique user id token. If this
is missing the function will attempt to connect to bpipe or sapi using the connection. It
is assumed that an app_name was set. See blpConnect() for app_name information}

\item{host}{A character variable with a hostname, defaults to 'localhost'}

\item{ip.address}{An optional character variable with an IP address
for authentication.  Usually the IP address where the uuid/user last
logged into the Bloomberg Terminal appication}

\item{con}{A connection object as created by a \code{blpConnect}
call, and retrieved via the internal function. This is the only required
argument to authenticate a bpipe connection with a appName.
\code{defaultConnection}.}

\item{default}{A logical indicating whether this authentication should
be saved as the default, as opposed to returned to the
user. Default to \code{TRUE}.}
}
\value{
In the \code{default=TRUE} case nothing is returned, and
this authentication is automatically used for all future calls which
omit the \code{identity} argument. Otherwise an authentication object is
returned which is required by all the accessor functions in the
package. (e.g. \code{bdp()} \code{bds()} \code{getPortfolio()}
}
\description{
This function authenticates against the the Bloomberg API
}
\examples{
\dontrun{
blpConnect(host=blpHost, port=blpPort)
blpAuthenticate(uuid=blpUUID, ip=blpIP_address)
bdp("IBM US Equity", "NAME")

blpid <- blpAuthenticate(uuid=blpUUID, ip=blpIP_address)
bdp("IBM US Equity", "NAME", identity=blpid)
}
}
\author{
Whit Armstrong and Dirk Eddelbuettel
}
