% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blpConnect.R
\name{blpConnect}
\alias{blpConnect}
\title{Establish connection to Bloomberg service}
\usage{
blpConnect(host = getOption("blpHost", "localhost"),
  port = getOption("blpPort", 8194L))
}
\arguments{
\item{host}{A character option with either a machine name that is resolvable by DNS, or an IP address. Defaults to \sQuote{localhost}.}

\item{port}{An integer variable with the connection port. Default to \code{8194L}.}
}
\value{
A connection object is returned. It is required by all the
accessor functions in th epackage.
}
\description{
This function connects to the Bloomberg API
}
\details{
For both \code{host} and \code{port} argument, default
values can also be specified via \code{\link{options}} using,
respectively, the named entries \code{blpHost} and
\code{blpConnect}.
}
\examples{
\dontrun{
  con <- blpConnect()   # adjust as needed
}
}
\author{
Whit Armstrong and Dirk Eddelbuettel
}

