% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blpDisconnect.R
\name{blpDisconnect}
\alias{blpDisconnect}
\title{Placeholder function for disconnection from Bloomberg}
\usage{
blpDisconnect(con)
}
\arguments{
\item{con}{A connection object}
}
\value{
A boolean is returned; it simply states whether the
connection object was small or large relative to an arbitrary
cutoff of 1000 bytes.
}
\description{
This function provides an empty stub and does not really disconnect.
}
\details{
The internal connection object is managed via
finalizers. As such the connection is only destroyed, and the
connection removed, once the packaged is unloaded or the session
is otherwise terminated.
}
\examples{
\dontrun{
  blpDisconnect(con)
}
}
\author{
Whit Armstrong and Dirk Eddelbuettel
}

