% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blpGenerateToken.R
\name{blpGenerateToken}
\alias{blpGenerateToken}
\title{Generate Bloomberg Security Token}
\usage{
blpGenerateToken(con = defaultConnection())
}
\arguments{
\item{con}{A connection object as created by a \code{blpConnect}
call. If not specified, the default connection is retrieved
via the internal function \code{defaultConnection}.}
}
\value{
A character vector containing the generated token.
}
\description{
This function generates a new token that can be used to further
authenticate against the Bloomberg API via B-PIPE.
}
\author{
Christopher Dudley
}
