% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defaultConnection.R
\name{defaultConnection}
\alias{defaultConnection}
\title{Return the default connection object}
\usage{
defaultConnection()
}
\description{
This function return the default connection object from the
package environment.  If no default connection has been
established yet, an error message is shown,
}
\details{
For the connection object, the required arguments
\code{host} and \code{port} argument can be set via
\code{\link{options}}. In addition, if an additional option
\code{blpAutoConnect} is set to \sQuote{TRUE}, a connection is
established in the \code{.onAttach()} function and stored in the
package environment. This effectively frees users from having to
explicitly create such an object. Of course, the user can also
call \code{blpConnect} explicitly and store the connection object.
This helper function looks up the stored connection object and
returns it. In case no connection has been established, and error
message is shown.
}
\examples{
\dontrun{
  con <- defaultConnection()
}
}
\author{
Whit Armstrong and Dirk Eddelbuettel
}

