% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.info.R
\name{field.info}
\alias{field.info}
\title{Run 'Bloomberg Field Data' Queries}
\usage{
field.info(fields, con = defaultConnection())
}
\arguments{
\item{fields}{A character vector with Bloomberg query fields.}

\item{con}{A connection object as created by a \code{blpConnect}
call, and retrieved via the internal function
\code{defaultConnection}.}
}
\value{
A data frame with as a many rows as entries in
\code{fields}
}
\description{
This function uses the Bloomberg API to retrieve 'bdp' (Bloomberg
Data Point) queries
}
\examples{
\dontrun{
  field.info(c("PX_LAST", "VOLUME"))
}
}
\author{
Whit Armstrong and Dirk Eddelbuettel
}

