% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fieldSearch}
\alias{fieldSearch}
\title{Search for matching data fields}
\usage{
fieldSearch(con, searchterm, excludeterm = "Static")
}
\arguments{
\item{con}{A connection object as return by a \code{blpConnect} call}

\item{searchterm}{A string with the term to search for}

\item{excludeterm}{A string with an expression for matches to excludes, defaults to \dQuote{Static}}
}
\value{
A \code{data.frame} with three columns of the id, mnenemonic and description of each match.
}
\description{
This function searches for matching Bloomberg data fields given a search term.
}
\examples{
\dontrun{
  con <- blpConnect()
  res <- fieldSearch(con, "vwap")
}
}
\author{
Dirk Eddelbuettel
}

