% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBars}
\alias{getBars}
\title{Get Open/High/Low/Close/Volume Bars from Bloomberg}
\usage{
getBars(con, security, eventType = "TRADE", barInterval = 60,
  startTime = Sys.time() - 60 * 60 * 6, endTime = Sys.time(),
  gapFillInitialBar = FALSE, verbose = FALSE, asXts = TRUE)
}
\arguments{
\item{con}{A connection object as return by a \code{blpConnect} call}

\item{security}{A character variable describing a valid security ticker}

\item{eventType}{A character variable describing an event type;
default is \sQuote{TRADE}}

\item{barInterval}{A integer denoting the number of minutes for each bar}

\item{startTime}{A Datetime object with the start time, defaults
to one hour before current time}

\item{endTime}{A Datetime object with the end time, defaults
to current time}

\item{gapFillInitialBar}{A boolean indicating whether the initial bar is to be filled, defaults to \sQuote{FALSE}}

\item{verbose}{A boolean indicating whether verbose operation is
desired, defaults to \sQuote{FALSE}}

\item{asXts}{A boolean indicating whether an \code{xts} object
should be returned, defaults to \sQuote{TRUE}.}
}
\value{
A numeric matrix with elements \sQuote{time},
\sQuote{open}, \sQuote{high}, \sQuote{low}, \sQuote{close},
\sQuote{numEvents}, \sQuote{volume}.  If \code{asXts} was set to
\sQuote{TRUE}, an \code{xts} object using the \sQuote{time} column
as index. Note that the \sQuote{time} value is adjusted: Bloomberg
returns the \emph{opening} time of the bar interval, whereas
financial studies typically refer to the most recent
timestamp. For this reason we add the length of the bar interval
to time value from Bloomberg to obtain the time at the end of the
interval.
}
\description{
This function uses the Bloomberg API to retrieve bars for the requested security.
}
\author{
Dirk Eddelbuettel
}

