% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getHeaderVersion}
\alias{getHeaderVersion}
\title{Get Bloomberg library header version}
\usage{
getHeaderVersion()
}
\value{
A string with four dot-separated values for major, minor,
pathch and build version of the headers.
}
\description{
This function retrieves the version of Bloomberg API headers.
}
\examples{
\dontrun{
   getHeaderVersion()
}
}
\seealso{
\code{getRuntimeVersion}
}
\author{
Dirk Eddelbuettel
}
