% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getRuntimeVersion}
\alias{getRuntimeVersion}
\title{Get Bloomberg library run-time version}
\usage{
getRuntimeVersion()
}
\value{
A string with four dot-separated values for major, minor,
pathch and build version of the run-time library.
}
\description{
This function retrieves the version of Bloomberg API run-time.
}
\examples{
\dontrun{
   getRuntimeVersion()
}
}
\author{
Dirk Eddelbuettel
}
\seealso{
\code{getHeaderVersion}
}

