% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTicks}
\alias{getTicks}
\title{Get Ticks from Bloomberg}
\usage{
getTicks(con, security, eventType = "TRADE", startTime = Sys.time() - 60 *
  60, endTime = Sys.time(), verbose = FALSE, asXts = TRUE)
}
\arguments{
\item{con}{A connection object as return by a \code{blpConnect} call}

\item{security}{A character variable describing a valid security ticker}

\item{eventType}{A character variable describing an event type;
default is \sQuote{TRADE}}

\item{startTime}{A Datetime object with the start time, defaults
to one hour before current time}

\item{endTime}{A Datetime object with the end time, defaults
to current time}

\item{verbose}{A boolean indicating whether verbose operation is
desired, defaults to \sQuote{FALSE}}

\item{asXts}{A boolean indicating whether an \code{xts} object
should be returned, defaults to \sQuote{TRUE}.}
}
\value{
A numeric matrix with elements \sQuote{time},
\sQuote{values} and \sQuote{sizes}.  If \code{asXts} was set to
\sQuote{TRUE}, an \code{xts} object using the \sQuote{time} column
as index.
}
\description{
This function uses the Bloomberg API to retrieve ticks for the requested security.
}
\author{
Dirk Eddelbuettel
}

