% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getVersionIdentifier}
\alias{getVersionIdentifier}
\title{Get Bloomberg library and run-time version}
\usage{
getVersionIdentifier()
}
\value{
A string (with unspecified format) containing the version
of the Blpapi runtime library.
}
\description{
This function retrieves the version string of the Bloomberg API.
}
\details{
Note that formatting of the returned string is described as
\sQuote{unspecified} by API documentation
}
\examples{
\dontrun{
   getVersionIdentifier()
}
}
\author{
Dirk Eddelbuettel
}
\seealso{
\code{getHeaderVersion}, \code{getRuntimeVersion}
}

