% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{haveBlp}
\alias{haveBlp}
\title{Get Blp availability in current build}
\usage{
haveBlp()
}
\value{
A logical indicating whether Blp was available at build.
}
\description{
This function returns a boolean indicating whether Blp support is available.
}
\author{
Dirk Eddelbuettel
}
