% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blpAuthenticate.R
\name{blpAuthenticate}
\alias{blpAuthenticate}
\title{Authenticate Bloomberg API access}
\usage{
blpAuthenticate(uuid, host = "localhost", ip.address, token,
  con = defaultConnection())
}
\arguments{
\item{uuid}{A character variable with a unique user id token}

\item{host}{A character variable with a hostname, defaults to 'localhost'}

\item{ip.address}{An optional character variable with an IP address}

\item{token}{An optional character variable with an authentication token}

\item{con}{A connection object as created by a \code{blpConnect}
call, and retrieved via the internal function
\code{defaultConnection}.}
}
\value{
An identity object that may be used to authorize further requests.
}
\description{
This function authenticates against the the Bloomberg API
}
\author{
Whit Armstrong and Dirk Eddelbuettel
}
