% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blpAuthenticate.R
\name{blpAuthenticate}
\alias{blpAuthenticate}
\title{Authenticate Bloomberg API access}
\usage{
blpAuthenticate(uuid, host = "localhost", ip.address,
  con = defaultConnection())
}
\arguments{
\item{uuid}{A character variable with a unique user id token}

\item{host}{A character variable with a hostname, defaults to 'localhost'}

\item{ip.address}{An optional character variable with an IP address}

\item{con}{A connection object as created by a \code{blpConnect}
call, and retrieved via the internal function
\code{defaultConnection}.}
}
\value{
Not sure. May just be the side effect of having the
session authenticated.
}
\description{
This function authenticates against the the Bloomberg API
}
\author{
Whit Armstrong and Dirk Eddelbuettel
}

