% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTicks}
\alias{getTicks}
\title{Get Ticks from Bloomberg}
\usage{
getTicks(con, security, eventType = "TRADE", startTime = Sys.time() - 60 *
  60, endTime = Sys.time(), verbose = FALSE,
  returnAs = getOption("blpType", "matrix"))
}
\arguments{
\item{con}{A connection object as return by a \code{blpConnect} call}

\item{security}{A character variable describing a valid security ticker}

\item{eventType}{A character variable describing an event type;
default is \sQuote{TRADE}}

\item{startTime}{A Datetime object with the start time, defaults
to one hour before current time}

\item{endTime}{A Datetime object with the end time, defaults
to current time}

\item{verbose}{A boolean indicating whether verbose operation is
desired, defaults to \sQuote{FALSE}}

\item{returnAs}{A character variable describing the type of return
object; currently supported are \sQuote{matrix} (also the default),
\sQuote{xts} and \sQuote{zoo}}
}
\value{
A numeric matrix with elements \sQuote{time},
\sQuote{values} and \sQuote{sizes}, or an object of type selected
in \code{returnAs}.
}
\description{
This function uses the Bloomberg API to retrieve ticks for the requested security.
}
\author{
Dirk Eddelbuettel
}

