% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTicks.R
\name{getTicks}
\alias{getTicks}
\title{Get Ticks from Bloomberg}
\usage{
getTicks(security, eventType = "TRADE", startTime = Sys.time() - 60 * 60,
  endTime = Sys.time(), verbose = FALSE, returnAs = getOption("blpType",
  "matrix"), setCondCodes = FALSE, tz = Sys.getenv("TZ", unset = "UTC"),
  con = .pkgenv$con)
}
\arguments{
\item{security}{A character variable describing a valid security ticker}

\item{eventType}{A character variable describing an event type. Multiple events can be specified
by a list \code{c("TRADE", "BEST_BID", "BEST_ASK")}, default is \sQuote{TRADE}}

\item{startTime}{A Datetime object with the start time, defaults
to one hour before current time}

\item{endTime}{A Datetime object with the end time, defaults
to current time}

\item{verbose}{A boolean indicating whether verbose operation is
desired, defaults to \sQuote{FALSE}}

\item{returnAs}{A character variable describing the type of return
object; currently supported are \sQuote{matrix} (also the default),
\sQuote{fts}, \sQuote{xts} and \sQuote{zoo}}

\item{setCondCodes}{A boolean indicating whether to return any ticks with condition codes,
associated with extraordinary trading and quoting circumstances,
defaults to \sQuote{FALSE}. Similar to setting \code{CondCodes = S} and \code{QRM = S} in Excel API}

\item{tz}{A character variable with the desired local timezone,
defaulting to the value \sQuote{TZ} environment variable, and
\sQuote{UTC} if unset}

\item{con}{A connection object as returned by a \code{blpConnect} call}
}
\value{
A numeric matrix with elements \sQuote{time}, (as a
\sQuote{POSIXct} object), \sQuote{values} and \sQuote{sizes}, or
an object of the type selected in \code{returnAs}.
}
\description{
This function uses the Bloomberg API to retrieve ticks for the requested security.
}
\author{
Dirk Eddelbuettel
}

